+++ test/offline/apps/fpm_demo_base/fpm_demo_base/fpm_demo_base/doctype/fpm_demo_note/fpm_demo_note.py
@@ -1,0 +1,8 @@
+# Copyright (c) 2026, FPM offline test
+# For license information, please see license.txt
+
+from frappe.model.document import Document
+
+
+class FPMDemoNote(Document):
+	pass